+++ packages/cubejs-playground/src/components/RollupDesigner/utils.ts
@@ -7,10 +7,12 @@ import {
 import { camelCase } from 'camel-case';
 
 import { QueryMemberKey } from '../../types';
+import { RollupSettings } from './components/Settings';
 
 export type PreAggregationReferences = {
   measures: string[];
   dimensions: string[];
+  segments: string[];
   timeDimensions: TimeDimensionBase[];
   timeDimension?: string;
   granularity?: TimeDimensionGranularity;
@@ -23,11 +25,13 @@ export type PreAggregationDefinition = {
 };
 
 export function getPreAggregationReferences(
-  transformedQuery: TransformedQuery | null
+  transformedQuery: TransformedQuery | null,
+  segments: Set<string>
 ): PreAggregationReferences {
   const references: PreAggregationReferences = {
     measures: [],
     dimensions: [],
+    segments: [],
     timeDimensions: [],
   };
 
@@ -36,7 +40,14 @@ export function getPreAggregationReferences(
   }
 
   if (transformedQuery?.sortedDimensions.length) {
-    references.dimensions = [...transformedQuery.sortedDimensions];
+    references.dimensions = [
+      ...transformedQuery.sortedDimensions.filter(
+        (name) => !segments.has(name)
+      ),
+    ];
+    references.segments = [
+      ...transformedQuery.sortedDimensions.filter((name) => segments.has(name)),
+    ];
   }
 
   if (
@@ -62,31 +73,32 @@ type PreAggregationDefinitionResult = {
 
 export function getPreAggregationDefinitionFromReferences(
   references: PreAggregationReferences,
-  name: string = 'main'
+  name: string = 'main',
+  settings: RollupSettings
 ): PreAggregationDefinitionResult {
-  const lines: string[] = [];
+  const { timeDimensions, ...otherReferences } = references;
+  const code: Record<string, any> = {
+    ...Object.entries(otherReferences).reduce(
+      (memo, [key, value]) => ({
+        ...memo,
+        ...(Array.isArray(value) && value.length ? { [key]: value } : null),
+      }),
+      {}
+    ),
+    ...settings,
+  };
 
-  if (references.measures.length) {
-    lines.push(`  measures: [${references.measures.map((m) => m).join(', ')}]`);
-  }
-
-  if (references.dimensions.length) {
-    lines.push(
-      `  dimensions: [${references.dimensions.map((m) => m).join(', ')}]`
-    );
-  }
-
-  if (references.timeDimensions.length) {
+  if (timeDimensions.length) {
     const { dimension, granularity } = references.timeDimensions[0];
 
-    lines.push(`  timeDimension: ${dimension}`);
+    code.timeDimension = dimension;
 
     if (granularity) {
-      lines.push(`  granularity: \`${granularity}\``);
+      code.granularity = `\`${granularity}\``;
     }
   }
 
-  const value = `{\n${lines.join(',\n')}\n}`;
+  const value = JSON.stringify(code, null, 2).replaceAll('"', '');
 
   return {
     code: `${camelCase(name)}: ${value}`,
@@ -107,7 +119,7 @@ export function updateQuery(
     } else {
       updatedQuery.timeDimensions = [
         {
-          // defafult granularity
+          // default granularity
           granularity: 'day',
           ...updatedQuery.timeDimensions?.[0],
           dimension: key,
@@ -125,4 +137,61 @@ export function updateQuery(
   }
 
   return updatedQuery;
+}
+
+function isBuffer(obj) {
+  return (
+    obj &&
+    obj.constructor &&
+    typeof obj.constructor.isBuffer === 'function' &&
+    obj.constructor.isBuffer(obj)
+  );
+}
+
+function keyIdentity(key) {
+  return key;
+}
+
+type FlattenOptions = {
+  safe?: boolean;
+  maxDepth?: number;
+  delimiter?: string;
+  transformKey?: (key: string) => string;
+};
+
+export function flatten(target: Object, opts: FlattenOptions = {}) {
+  const delimiter = opts.delimiter || '.';
+  const maxDepth = opts.maxDepth || 100;
+  const transformKey = opts.transformKey || keyIdentity;
+  const output = {};
+
+  function step(object, prev = null, keyDepth = 1) {
+    Object.keys(object).forEach(function (key) {
+      const value = object[key];
+      const isArray = opts.safe && Array.isArray(value);
+      const type = Object.prototype.toString.call(value);
+      const isbuffer = isBuffer(value);
+      const isobject = type === '[object Object]' || type === '[object Array]';
+
+      const newKey = prev
+        ? prev + delimiter + transformKey(key)
+        : transformKey(key);
+
+      if (
+        !isArray &&
+        !isbuffer &&
+        isobject &&
+        Object.keys(value).length &&
+        (!opts.maxDepth || keyDepth < maxDepth)
+      ) {
+        return step(value, newKey, keyDepth + 1);
+      }
+
+      output[newKey] = value;
+    });
+  }
+
+  step(target);
+
+  return output;
 }