+++ packages/cubejs-server-core/src/core/RefreshScheduler.ts
@@ -51,25 +51,28 @@ export class RefreshScheduler {
     const orchestratorApi = this.serverCore.getOrchestratorApi(context);
     const preAggregationsLoadCacheByDataSource = {};
 
+    // Return a empty array for cases with 2 same pre-aggreagtions but with different partitionGranularity
+    // Only the most detailed pre-aggregations will be use
+    if (!preAggregationDescription) {
+      return [];
+    }
+
     const partitions = await orchestratorApi.expandPartitionsInPreAggregations({
       preAggregations: [preAggregationDescription],
       preAggregationsLoadCacheByDataSource,
       requestId: context.requestId
     });
 
-    return Promise.all(partitions.preAggregations.map(async partition => {
-      delete partition.preAggregationStartEndQueries;
-      return {
-        query: {
-          ...baseQuery,
-          timeDimensions: baseQuery.timeDimensions && baseQuery.timeDimensions[0] && [{
-            ...baseQuery.timeDimensions[0],
-            dateRange: partition.loadSql[1]
-          }]
-        },
-        sql: partition
-      };
-    }));
+    return Promise.all(partitions.preAggregations.map(async partition => ({
+      query: {
+        ...baseQuery,
+        timeDimensions: baseQuery.timeDimensions && baseQuery.timeDimensions[0] && [{
+          ...baseQuery.timeDimensions[0],
+          dateRange: partition.loadSql[1]
+        }]
+      },
+      sql: partition
+    })));
   }
 
   protected async baseQueryForPreAggregation(